done education-detail page
